requirements.txt
Flask==2.3.3
firebase-admin==6.2.0
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
opencv-python==4.8.0.74
face-recognition==1.3.0
SpeechRecognition==3.10.0
# pyaudio==0.2.13
pyserial==3.5
mfrc522==0.0.7 